# Conflicts:
#	fighterCardGame/fighterCardGame/target/classes/com/kurapati/fighterCardGame/auctionHouse/AuctionHouseCard.class
#	fighterCardGame/fighterCardGame/target/classes/com/kurapati/fighterCardGame/auctionHouse/AuctionHouseCardService.class
#	fighterCardGame/fighterCardGame/target/classes/com/kurapati/fighterCardGame/auctionHouse/AuctionHouseController.class
#	fighterCardGame/fighterCardGame/target/classes/com/kurapati/fighterCardGame/pack/PackService.class
#	fighterCardGame/fighterCardGame/target/classes/com/kurapati/fighterCardGame/users/UserController.class
#	fighterCardGame/fighterCardGame/target/classes/com/kurapati/fighterCardGame/users/Users.class

+++ fighterCardGame/fighterCardGame/src/main/java/com/kurapati/fighterCardGame/auctionHouse/AuctionHouseCardService.java
@@ -5,11 +5,13 @@ import com.kurapati.fighterCardGame.card.CardRepository;
 import com.kurapati.fighterCardGame.users.UserRepository;
 import com.kurapati.fighterCardGame.users.Users;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
+@Transactional
 public class AuctionHouseCardService {
 
     private final AuctionHouseCardRepository auctionHouseCardRepo;
@@ -23,11 +25,11 @@ public class AuctionHouseCardService {
     }
 
     public void createAuctionHouseCard(int sellerId, int cardId, int auctionStartingPrice, int buyNowPrice, int duration){
+        Users seller = userRepository.findById(sellerId).orElseThrow(
+                ()->new RuntimeException("user does not exist"));
         auctionHouseCardRepo.save(
                 new AuctionHouseCard(
-                        userRepository.findById(sellerId).orElseThrow(
-                                ()->new RuntimeException("user does not exist")
-                        ),
+                        seller,
                         cardRepo.findById(cardId).orElseThrow(
                                 ()->new RuntimeException("Card does not exist")
                         ),
@@ -36,45 +38,58 @@ public class AuctionHouseCardService {
                         duration
                 )
         );
+        seller.setUserCards(
+                seller.getUserCards().stream()
+                        .filter(sellerCard -> sellerCard.getCardId()!=cardId)
+                        .collect(Collectors.toList())
+        );
 
     }
 
     public void buyAuctionHouseCard(int auctionCardId,int sellerId, int buyerId, int cardId, int sellingPrice){
+
         //remove card from seller
         Users seller = userRepository.findById(sellerId).orElseThrow(
                 ()->new RuntimeException("user does not exist")
         );
-        seller.setUserCards( seller.getUserCards().stream()
-                .filter(card -> card.getCardId()!=cardId)
-                .collect(Collectors.toList())
-        );
-        //add money to buyer
-        seller.setMoney(seller.getMoney()+sellingPrice);
-
-        //add card to buyer
         Users buyer = userRepository.findById(buyerId).orElseThrow(
                 ()->new RuntimeException("user does not exist")
         );
         Card card = cardRepo.findById(cardId).orElseThrow(
-                ()->new RuntimeException("user does not exist")
+                ()->new RuntimeException("card does not exist")
         );
+        if(buyer.getMoney() > sellingPrice){
+            seller.setUserCards( seller.getUserCards().stream()
+                    .filter(sellerCard -> sellerCard.getCardId()!=cardId)
+                    .collect(Collectors.toList())
+            );
+            //add money to buyer
+            seller.setMoney(seller.getMoney()+sellingPrice);
 
-        List<Card> cards =buyer.getUserCards();
-        cards.add(card);
-        buyer.setUserCards(cards);
-        //remove money form buyer
-        buyer.setMoney(buyer.getMoney()-sellingPrice);
-        //remove card from auctionHouse
-        auctionHouseCardRepo.deleteById(auctionCardId);
+            //add card to buyer
+
+            List<Card> cards =buyer.getUserCards();
+            cards.add(card);
+            buyer.setUserCards(cards);
+            //remove money form buyer
+            buyer.setMoney(buyer.getMoney()-sellingPrice);
+            //remove card from auctionHouse
+            auctionHouseCardRepo.deleteById(auctionCardId);
+        }
+
     }
 
     public void placeBid(int bidderId, int auctionCardId,int bidAmount){
-        AuctionHouseCard auctionHouseCard = auctionHouseCardRepo.findById(auctionCardId).orElseThrow();
-        auctionHouseCard.setMaxBidder(userRepository.findById(bidderId).orElseThrow());
-        auctionHouseCard.setStartingAuction(bidAmount);
-
         Users bidder = userRepository.findById(bidderId).orElseThrow();
-        bidder.setMoney(bidder.getMoney()-bidAmount);
+        if(bidder.getMoney() >= bidAmount) {
+            System.out.println(auctionCardId);
+            AuctionHouseCard auctionHouseCard = auctionHouseCardRepo.findById(auctionCardId).orElseThrow();
+            auctionHouseCard.setMaxBidder(userRepository.findById(bidderId).orElseThrow());
+            auctionHouseCard.setBidAmount(bidAmount);
+
+
+            bidder.setMoney(bidder.getMoney() - bidAmount);
+        }
     }
 
     public void updateAuctionHouseCard(){
